Removed profile.service and moved the functions in post.service and post.ts
